Deduplicate RunSettings providers via shared base classes
Extract the shared logic of the three MSTest-native RunSettings providers
and their VSTestBridge counterparts into abstract base classes in
SharedExtensionHelpers (RunSettingsCommandLineOptionsProviderBase,
RunSettingsConfigurationProviderBase, RunSettingsEnvironmentVariableProviderBase),
linked into both assemblies. The concrete providers now only supply the
per-package resource strings, identity, and runsettings-reading behavior,
eliminating ~180 lines of structural duplication.

Fixes #9875

Copilot-Session: 6651b296-3200-41b4-b72a-b9c35e648bcf
