+++ src/Platform/Microsoft.Testing.Extensions.VSTestBridge/TestHostControllers/RunSettingsEnvironmentVariableProvider.cs
@@ -1,55 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Testing.Extensions.VSTestBridge.CommandLine;
 using Microsoft.Testing.Platform.CommandLine;
 using Microsoft.Testing.Platform.Extensions;
-using Microsoft.Testing.Platform.Extensions.TestHostControllers;
 using Microsoft.Testing.Platform.Helpers;
 
 namespace Microsoft.Testing.Extensions.VSTestBridge.TestHostControllers;
 
-internal sealed class RunSettingsEnvironmentVariableProvider : ITestHostEnvironmentVariableProvider
-{
-    private readonly IExtension _extension;
-    private readonly ICommandLineOptions _commandLineOptions;
-    private readonly IFileSystem _fileSystem;
-    private readonly IEnvironment _environment;
-    private XDocument? _runSettings;
-
-    public RunSettingsEnvironmentVariableProvider(IExtension extension, ICommandLineOptions commandLineOptions, IFileSystem fileSystem, IEnvironment environment)
-    {
-        _extension = extension;
-        _commandLineOptions = commandLineOptions;
-        _fileSystem = fileSystem;
-        _environment = environment;
-    }
-
-    public string Uid => _extension.Uid;
-
-    public string Version => _extension.Version;
-
-    public string DisplayName => _extension.DisplayName;
-
-    public string Description => _extension.Description;
-
-    public async Task<bool> IsEnabledAsync()
-    {
-        _runSettings = await RunSettingsProviderHelper.TryLoadRunSettingsAsync(
-            _commandLineOptions,
-            _fileSystem,
-            _environment,
-            RunSettingsCommandLineOptionsProvider.RunSettingsOptionName).ConfigureAwait(false);
-
-        return _runSettings is not null && RunSettingsProviderHelper.HasEnvironmentVariables(_runSettings);
-    }
-
-    public Task UpdateAsync(IEnvironmentVariables environmentVariables)
-    {
-        RunSettingsProviderHelper.ApplyEnvironmentVariables(_runSettings!, environmentVariables);
-        return Task.CompletedTask;
-    }
-
-    public Task<ValidationResult> ValidateTestHostEnvironmentVariablesAsync(IReadOnlyEnvironmentVariables environmentVariables)
-        => ValidationResult.ValidTask;
-}
+internal sealed class RunSettingsEnvironmentVariableProvider(IExtension extension, ICommandLineOptions commandLineOptions, IFileSystem fileSystem, IEnvironment environment)
+    : RunSettingsEnvironmentVariableProviderBase(extension, commandLineOptions, fileSystem, environment);
